feat(formal): extraction P4 — exact equipment scoring; fifteen extracted cores; closure completeness universal
P4a: the last floats in the equipment-decision path replaced with exact
arithmetic. equip_value/tool_value are int (every summand always was);
the whole strategy ranking pipeline (priors, gear gains, blend weights,
decide keys, objective gap fractions, personality weights) moves to
Fraction; learned float yields lift exactly once at the boundary
(the P3c idiom). Diff oracles compare production values to the Lean Rat
models with zero tolerance — the float-tolerance seams are gone from
equipment scoring. Honesty note honored: behavior is deliberately not
byte-identical; a 200k-pair sweep found ZERO true order inversions and
92 float-"strict" orderings that were rounding noise now resolved as
exact ties falling to documented tiebreaks (witness: 2/5*1/5*3/2 vs
3/10*1/5*2, both exactly 3/25, where float said one was greater by
2e-17). The float -inf sentinel in upgrade comparison retired in favor
of no-candidate semantics.

P4b: equipment/scoring.py and tiers/equip_value.py hoisted to pure
cores and extracted (registry: 15 modules, zero new extractor
constructs needed). Bridges7.lean: all bridges FULL universal —
weapon/armor scores over an injective element embedding, equip_value
pointwise; clamp, strictness, tiebreak, no-downgrade, and
gather-optimality theorems transferred onto the extracted defs.
tool_value bridged by its disclosed duality with gather_score.

P4c: recipe_closure completeness upgraded from kernel pins to a
universal theorem. The fuel measure (count of unmarked recipe entries)
strictly decreases down every recursion path, so the len(recipes)+1
seed never exhausts; recipe_closure_pure_spec is now a complete iff —
output membership <-> least-fixpoint reachability, every graph, no
wellformedness hypothesis (entry counting handles duplicate keys).
Six completeness pins dropped as superseded; the two numeric pins the
mutation suite cites remain.

P4 closes the deferred ledger: every pure decision core in the
codebase now has a mechanically extracted, kernel-bridged Lean model.

+++ src/artifactsmmo_cli/ai/goals/upgrade_selection.py
@@ -28,7 +28,7 @@ class UpgradeCandidate:
     """A single equipment-upgrade pick, pre-resolved from game_data."""
 
     item_code: str
-    value: float        # equip_value (attack + resistance + hp_restore; integer-valued)
+    value: int          # equip_value (exact int since P4a; was float-typed before)
     level: int
     craft_level: int
     relevant: bool      # relevant tool/skill for the active task
@@ -49,13 +49,13 @@ def best_by_value(inv: UpgradeCandidate | None,
     return inv if inv.value >= craft.value else craft
 
 
-def craftable_key(c: UpgradeCandidate) -> tuple[int, int, float, int, str]:
+def craftable_key(c: UpgradeCandidate) -> tuple[int, int, int, int, str]:
     """Lexicographic sort key for craftable candidates:
     (relevant, fills_empty, value, -craft_level, item_code)."""
     return (int(c.relevant), int(c.fills_empty), c.value, -c.craft_level, c.item_code)
 
 
-def inventory_key(c: UpgradeCandidate) -> tuple[int, float, int, str]:
+def inventory_key(c: UpgradeCandidate) -> tuple[int, int, int, str]:
     """Lexicographic sort key for inventory candidates:
     (relevant, value, level, item_code)."""
     return (int(c.relevant), c.value, c.level, c.item_code)
